+++ configs/default.yaml
@@ -166,8 +166,8 @@ model:
       in_channels: 12
       d_model: 256
       nhead: 8
-      num_encoder_layers: 4
-      num_decoder_layers: 4
+      num_encoder_layers: 6
+      num_decoder_layers: 6
       dim_feedforward: 1024
       dropout: 0.1
